% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection.multiple}
\alias{bisection.multiple}
\title{bisection.multiple}
\usage{
bisection.multiple(
  A,
  B,
  lambda = NULL,
  nv = 2,
  max_iter = 1e+05,
  tol = 1e-06,
  ...
)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{list of background covariance(s).}

\item{lambda}{initial guess on what the Lagrange Multiplier should be. Default NULL}

\item{nv}{number of uca components to estimate}

\item{max_iter}{maximum iterations for coordinate descent, if tolerance is not reached. default 1E5}

\item{tol}{tolerance for stopping criteria of coordinate descent. default 1E-6}

\item{...}{other parameters to pass in to bisection(...)}
}
\value{
for each background covariance matrix in B, return list of tau (the optimal lagrange multiplier), vectors(eigenvector)
 associated with largest value (eigenvalue)
}
\description{
bisection method but for multiple background datasets. Use bisection method to find the optimal Lagrangian for solving discriminant PCA of each background dataset. depends on irlba
}
