% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{bisection}
\alias{bisection}
\title{bisection}
\usage{
bisection(A, B, limit = c(0, 20), maxit = 1e+05, nv = 1, tol = 1e-06)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{Background Covariance Matrix.}

\item{limit}{a vector of c(lower, upper) bounds}

\item{maxit}{maxium number of iterations for the algorithm to run}

\item{nv}{number of eigenvectors to use}

\item{tol}{tolerance for when to stop the algorithm}
}
\value{
the final list of tau (the optimal lagrange multiplier), vector (eigenvector)
 associated with tau, value (eigenvalue), and score (derivative value of the lagrangian)
}
\description{
Use bisection method to find the optimal Lagrangian for solving unique component analysis (uca)
}
