% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eigen.R
\name{bisection2}
\alias{bisection2}
\title{bisection2}
\usage{
bisection2(A, B, limit = c(0, 20), maxit = 100000L, tol = 1e-06)
}
\arguments{
\item{A}{a n1xp data matrix}

\item{B}{a n2xp data matrix}

\item{limit}{bounds for the lagrange multiplier.}

\item{maxit}{maximum iterations}

\item{tol}{tolerance for convergence criteria}
}
\value{
list of tau, largest eigenvalue, and score
}
\description{
compute the UCA for single background using SVD and QR. good for bigger data where loading the covariance matrix is difficult
}
