% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_bisection_methods.R
\name{bisection_cov}
\alias{bisection_cov}
\title{Use bisection to find the optimal lagrange multiplier for a single background}
\usage{
bisection_cov(A, B, limit = 50, maxit = 100000L, nv = 1, tol = 1e-06)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{Background Covariance Matrix.}

\item{limit}{upperbound of lagrange multiplier}

\item{maxit}{maxium number of iterations for the algorithm to run}

\item{nv}{number of eigenvectors to use}

\item{tol}{tolerance for when to stop the algorithm}
}
\value{
list of two elements:
\itemize{
 \item tau: the lagrange multiplier
 \item score: eigenvalue associated with tau
}
}
\description{
Use bisection method to find the optimal Lagrange multiplier when provided
covariance matrices and a single background matrix
}
