% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bisection_methods.R
\name{bisection_data_multiple}
\alias{bisection_data_multiple}
\title{Use bisection to find the optimal lagrange multiplier for multiple background
without constructing covariance matrices}
\usage{
bisection_data_multiple(
  B_focus,
  t_A,
  t_B,
  right,
  svd_right,
  lambda,
  j,
  limit = 20L,
  maxit = 1e+05,
  tol = 1e-06
)
}
\arguments{
\item{t_A}{precomputed A transpose}

\item{t_B}{precomputed B transpose}

\item{right}{precomputed right long matrix: rbind(A, B)}

\item{svd_right}{precomputed svd of right matrix}

\item{lambda}{j dimensional vector with candidate lagrange multipliers for
each background data matrix}

\item{j}{index of specific background you're solving for}

\item{limit}{upperbound for the lagrange multiplier}

\item{maxit}{maximum iterations}

\item{tol}{tolerance for convergence criteria}

\item{B_focusn}{x p data matrix of background matrix to solve lagrangian for}
}
\value{
list of two elements:
\itemize{
 \item tau: the lagrange multiplier
 \item score: eigenvalue associated with tau
 }
}
\description{
Use bisection method to find the optimal Lagrange multiplier using data
matrices and multiple background matrices. This is useful when constructing
the covariance matrix is computationally intensive.
}
