% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eigen.R
\name{broken_svd}
\alias{broken_svd}
\title{broken_svd}
\usage{
broken_svd(left, right, nv)
}
\arguments{
\item{left}{left side of a product}

\item{right}{right side of a product}

\item{nv}{number of unique components}
}
\value{
top nv eigenvalues and associated eigenvectors
}
\description{
calculate SVD of a product of matrices by using svd and QR decompositions
}
