% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eigen.R
\name{center}
\alias{center}
\title{centeer}
\usage{
center(X)
}
\arguments{
\item{X}{a matrix}
}
\value{
centered data matrix X
}
\description{
Convenient function to center the data, rather than typing `scale' or `sweep`
}
