% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bisection_methods.R
\name{data_multiple}
\alias{data_multiple}
\title{UCA for multiple backgrounds using data matrices}
\usage{
data_multiple(
  A,
  B,
  lambda = NULL,
  nv = 2L,
  max_iter = 100000L,
  tol = 1e-06,
  algo = "bisection",
  ...
)
}
\arguments{
\item{A}{Target Data Matrix. n1 x p dimensions}

\item{B}{List of k Background Data Matrix. n_k x p dimensions}

\item{lambda}{initial guess for Lagrange Multiplier. Default NULL.
algo = "bisection" only}

\item{nv}{number of uca components to estimate}

\item{max_iter}{maximum iterations for all algorithms if tolerance is not
reached. default 1E5.}

\item{tol}{tolerance for stopping criteria for all algorithms. default 1E-6}

\item{algo}{which algorithm to use. default algo == "bisection". If
algo = "cd", L-BFGS-B optimization is used instead for coordinate descent.}

\item{...}{additional parameters to pass in to `bisection_data_multiple()`
and `optim_data_cd()`}
}
\value{
list of three elements:
\itemize{
 \item tau: the optimal lagrange multiplier(s)
 \item values: eigenvalue associated with tau(s)
 \item vectors: top `nv` eigenvectors associated with tau(s)
 }
}
\description{
Wrapper function for UCA with multiple background datasets with data
matrices. We use a sequence of SVD and QR decomposition (Product-SVD).
Use different algorithms to find the optimal Lagrangian
}
