% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{gr_fn_multiple}
\alias{gr_fn_multiple}
\title{gr_fn_multiple}
\usage{
gr_fn_multiple(lambda, B, t_A, t_B, lambda_B, right, svd_right, j)
}
\arguments{
\item{lambda}{the lagrange multiplier}

\item{B}{background interested in calculating lagrange multiplier for}

\item{t_A}{transpose of A, a pxn1 data matrx}

\item{t_B}{list of transpose of B, a pxn2 data matrix}

\item{right}{rbind(A, B), a (n1+n2)xp data matrix}

\item{svd_right}{svd of right data matrix object}

\item{j}{which background matrix are we focusing on}
}
\value{
value of the gradient fn
}
\description{
gradient function multiple backgroundsimplementation, data method
}
