% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eigen.R
\name{magic_eigen_multiple}
\alias{magic_eigen_multiple}
\title{magic_eigen_multiple}
\usage{
magic_eigen_multiple(
  B_focus,
  t_A,
  t_B,
  right,
  svd_right,
  lambda,
  j,
  limit = 20L,
  maxit = 1e+05,
  tol = 1e-06
)
}
\arguments{
\item{B_focus}{a nxp data matrix of the background we're solving lagrangian for}

\item{t_A}{precomputed A transpose}

\item{t_B}{precomputed B transpose}

\item{right}{precomputed right long matrix: rbind(A, B)}

\item{svd_right}{precomputed svd of right matrix}

\item{lambda}{a j dimensional vector with possible lagrange multipliers for each background data matrix}

\item{j}{the specific background you're solving for}

\item{limit}{upperbound for the lagrange multiplier.}

\item{maxit}{maximum iterations}

\item{tol}{tolerance for convergence criteria}
}
\value{
list of tau, largest eigenvalue, and score
}
\description{
Solve for the optimal lagrange multiplier for the jth background. used only when multiple backgrounds exist.
}
