% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eigen.R
\name{multiple_score_calc}
\alias{multiple_score_calc}
\title{multiple_score_calc}
\usage{
multiple_score_calc(left, right, svd_right, tau, B_focus)
}
\arguments{
\item{left}{a rectangular matrix with p columns}

\item{right}{a rectangular matrix with p rows}

\item{svd_right}{a precalculated svd for the right matrix. saves time}

\item{tau}{contrastive parameter}
}
\value{
list of the largest eigenvalue, associated score, and tau
}
\description{
calculate the eigendecomposition via a product of matrices using SVD and QR decomposition. Much faster for large matrices.
 compute the lambda for the jth background
}
