% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_optim_methods.R
\name{obj_fun_multiple_gd}
\alias{obj_fun_multiple_gd}
\title{objective function for gradient descent implementation in multi-background
with covariance matrices}
\usage{
obj_fun_multiple_gd(A, B_unlist, tau)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B_unlist}{B unlist of background covariance(s).}

\item{tau}{lowerbound for root finding}
}
\value{
value of objective function
}
\description{
objective function for gradient descent implementation in multi-background
with covariance matrices
}
