% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim_bfgs}
\alias{optim_bfgs}
\title{optim_bfgs implementaiton, cov method}
\usage{
optim_bfgs(A, B, maxit = 500L, nv = 1)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{Background Covariance Matrix.}

\item{maxit}{maxium number of iterations for the algorithm to run}

\item{nv}{number of eigenvectors to use}
}
\value{
the final list of tau (the optimal lagrange multiplier), vector (eigenvector)
 associated with tau, value (eigenvalue), and score (derivative value of the lagrangian)
}
\description{
Use optim-L-BFGS-S to find the optimal Lagrangian for solving unique component analysis (uca)
}
