% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim_bfgs2}
\alias{optim_bfgs2}
\title{optim_bfgs2: optim implementation for the data method}
\usage{
optim_bfgs2(A, B, maxit = 500L, nv = 1)
}
\arguments{
\item{A}{a n1xp data matrix}

\item{B}{a n2xp data matrix}

\item{maxit}{maximum iterations}
}
\value{
list of tau, largest eigenvalue, and score
}
\description{
compute the UCA for single background using SVD and QR. good for bigger data where loading the covariance matrix is difficult using optim then bisection
}
