% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim_bfgs_multiple}
\alias{optim_bfgs_multiple}
\title{optim_bfgs_multiple}
\usage{
optim_bfgs_multiple(
  B_focus,
  t_A,
  t_B,
  right,
  svd_right,
  lambda,
  j,
  maxit = 500
)
}
\arguments{
\item{B_focus}{single background interested in calculating lagrange multiplier for}

\item{t_A}{transpose of A, a pxn1 data matrx}

\item{t_B}{list of transpose of B, a pxn2 data matrix}

\item{right}{rbind(A, B), a (n1+n2)xp data matrix}

\item{svd_right}{svd of right data matrix object}

\item{lambda}{the lagrange multiplier}
}
\value{
optimum eigenvalue and lagrange multiplier for a single background in the multi. background setting
}
\description{
objective function multiple backgroundsimplementation, data method
}
