% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optim_methods.R
\name{optim_cd_multiple}
\alias{optim_cd_multiple}
\title{coordinate descent BFGS implementation in multi-background setting
without covariance matrix}
\usage{
optim_cd_multiple(B_focus, t_A, t_B, right, svd_right, lambda, j, maxit = 500)
}
\arguments{
\item{B_focus}{background interested in calculating lagrange multiplier for}

\item{t_A}{transpose of A, a p x n1 data matrx}

\item{t_B}{list of transpose of B, a p x n2 data matrix}

\item{right}{rbind(A, B), a (n1 + n2) x p data matrix}

\item{svd_right}{svd of right data matrix object}

\item{lambda}{the lagrange multiplier}
}
\value{
list of two elements:
\itemize{
 \item values: optimal eigenvalue
 \item tau: the contrastive parameter (lagrange multiplier) for single background
 in the multi. background setting
 }
}
\description{
Compute the UCA lagrange multiplier for a single background in the multiple
backgrounds using SVD and QR in a coordinate descent manner using optim.
Faster for high-dimensional data since we avoid creating a covariance matrix
}
