% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_optim_methods.R
\name{optim_cov_bfgs_gd}
\alias{optim_cov_bfgs_gd}
\title{Gradient Descent with BFGS implementation for single OR multi-background
with covariance matrices}
\usage{
optim_cov_bfgs_gd(A, B, maxit = 500L, nv = 1)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{List of background Covariance Matrices}

\item{maxit}{maxium number of iterations for the algorithm to run}

\item{nv}{number of eigenvectors to use}
}
\value{
list of two elements:
\itemize{
 \item values: optimal eigenvalue
 \item tau: the contrastive parameter(s) (lagrange multiplier(s)) for each
 background in the multi. background setting
 }
}
\description{
Use L-BFGS-B to find the optimal Lagrange Multipliers for the optimal
Lagrangian for a single, or multiple backgrounds, simulatneously, using
covariance matrices.
}
