% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_optim_methods.R
\name{optim_cov_cd}
\alias{optim_cov_cd}
\title{BFGS Implementation using covariance matrices for single background}
\usage{
optim_cov_cd(A, B, maxit = 500L, nv = 1)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{Background Covariance Matrix.}

\item{maxit}{maxium number of iterations for the algorithm to run}

\item{nv}{number of eigenvectors to use}
}
\value{
list of two elements:
\itemize{
 \item values: optimal eigenvalue
 \item tau: the contrastive parameter (lagrange multiplier)
 }
}
\description{
Compute the UCA dimensions for single background using optim-L-BFGS-B to
find the optimal Lagrange multiplier
}
