% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_optim_methods.R
\name{optim_data_cd}
\alias{optim_data_cd}
\title{BFGS implementation for single background using the data method
(without using covariance matrix)}
\usage{
optim_data_cd(A, B, maxit = 500L)
}
\arguments{
\item{A}{a n1 x p data matrix}

\item{B}{a n2 x p data matrix}

\item{maxit}{maximum iterations}
}
\value{
list of two elements:
\itemize{
 \item values: eigenvalues
 \item tau: the contrastive parameter
 }
}
\description{
Compute the UCA dimensions for single background using SVD and QR
using optim. Faster for high-dimensional data because we avoid
constructing the covariance matrix
}
