% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{product_svd_R}
\alias{product_svd_R}
\title{SVD of a difference of covariance matrices without constructing cov matrix}
\usage{
product_svd_R(left, right, nv)
}
\arguments{
\item{left}{left side of a product}

\item{right}{right side of a product}

\item{nv}{number of unique components}
}
\value{
top nv eigenvalues and associated eigenvectors
}
\description{
calculate SVD of a product of matrices by using svd and QR decompositions
}
