% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_bisection_methods.R
\name{score_calc}
\alias{score_calc}
\title{Calculate the values of the Lagrangian}
\usage{
score_calc(A, B, tau)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{list of background covariance(s).}

\item{tau}{lowerbound for root finding}
}
\value{
list of three elements:
\itemize{
 \item tau, the lagrange multiplier,
 \item eigenvalues associated with tau
 \item derivative value of the lagrangian
 }
}
\description{
Return the value and derivative of the Lagrangian
}
