% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uca.R
\docType{package}
\name{uca-package}
\alias{uca-package}
\alias{_PACKAGE}
\title{uca: Unique Component Analysis}
\description{
Unique Component Analysis (UCA) looks for directions which maximize the covariance in a target dataset subject to minimizing the covariance in one or more background data set. This package allows you to run UCA with either the original data matrix or the covariance matrix.
}
