% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uca.R
\name{uca}
\alias{uca}
\title{Unique Component Analysis}
\usage{
uca(
  A,
  B,
  nv = 2,
  method = "data",
  center = FALSE,
  scale = FALSE,
  algo = "cd",
  ...
)
}
\arguments{
\item{A}{Target Data or Covariance Matrix}

\item{B}{list of background data or covariance matrices.}

\item{nv}{number of uca components to estimate}

\item{method}{method used to calculate the uca values and vectors.
Use method = 'data' when passing a n x p data matrix. Recommended when p > n
Use method = 'cov' when passing in a covariance matrix.}

\item{center}{logical: default FALSE. If TRUE, data matrix A & B are centered}

\item{scale}{logical: default FALSE. If TRUE, data matrix A & B are centered
and scaled, regardless of whether `center == TRUE` or `center == FALSE`}

\item{algo}{algorithm to find lagrange multiplier(s). valid algorithms are
"bisection", "cd" (coordinate descent), and "gd" (gradient descent). For
single background data, "cd" and "gd" are the same. default is "cd", but
bisection exists for backwards compatibility.}
}
\value{
list of three elements:
\itemize{
 \item values (eigenvalues)
 \item vectors (eigenvectors) of unique component analysis
 \item tau optimal Lagrange Multiplier(s) associated with uca
 }
}
\description{
Run unique component analysis
}
