% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eigen.R
\name{bisection2.multiple}
\alias{bisection2.multiple}
\title{bisection2.multiple}
\usage{
bisection2.multiple(
  A,
  B,
  lambda = NULL,
  nv = 2L,
  max_iter = 100000L,
  tol = 1e-06,
  ...
)
}
\arguments{
\item{A}{*Centered* Target Data Matrix. n1 x p dimensions}

\item{B}{*Centered* List of k Background Data Matrix. n_k x p dimensions}

\item{lambda}{contrastive parameters if known. used to start algorithm. defaults to NULL}

\item{nv}{number of uca components to calculate}

\item{max_iter}{maximum number of iterations before giving up}

\item{tol}{convergence criteria for coordinate descent}
}
\value{
list of tau, largest eigenvalue, and score
}
\description{
Solving for the unique components using SVD and QR in the instance of multiple backgrounds.
}
