% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_optim_methods.R
\name{gr_fun}
\alias{gr_fun}
\title{gradient function for BFGS implementation using covariance matrices}
\usage{
gr_fun(A, B, tau)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{list of background covariance(s).}

\item{tau}{lowerbound for root finding}
}
\value{
value of gradient function
}
\description{
gradient function for BFGS implementation using covariance matrices
}
