% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magic_eigen.R
\name{magic_eigen_multiple}
\alias{magic_eigen_multiple}
\title{magic_eigen_multiple}
\usage{
magic_eigen_multiple(
  A,
  B,
  lambda,
  j,
  limit = c(0, 20),
  maxit = 1e+05,
  tol = 1e-06
)
}
\arguments{
\item{A}{a *Centered* n1xp data matrix}

\item{B}{*Centered* a n2xp data matrix}

\item{lambda}{a j dimensional vector with possible lagrange multipliers for each background data matrix}

\item{j}{the specific background you're solving for}

\item{limit}{bounds for the lagrange multiplier.}

\item{maxit}{maximum iterations}

\item{tol}{tolerance for convergence criteria}
}
\value{
list of tau, largest eigenvalue, and score
}
\description{
Solve for the optimal lagrange multiplier for the jth background. used only when multiple backgrounds exist.
}
