% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{uca}
\alias{uca}
\title{uca}
\usage{
uca(A, B, nv = 2, ...)
}
\arguments{
\item{A}{Target Covariance Matrix}

\item{B}{list of background covariance(s).}

\item{nv}{number of uca components to estimate}

\item{...}{other parameters to pass in to bisection(...) and bisection.multiple(...)}
}
\value{
values(eigenvalues), vectors (eigenvectors), tau (Lagrange Multiplier) associated with unique component analysis
}
\description{
Run unique component analysis
}
