% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection.R
\name{uca}
\alias{uca}
\title{uca}
\usage{
uca(A, B, nv = 2, method = "data", center = F, scale = T, ...)
}
\arguments{
\item{A}{Target Data or Covariance Matrix}

\item{B}{list of background data or covariance matrices.}

\item{nv}{number of uca components to estimate}

\item{method}{method used to calculate the uca values and vectors.}

\item{center}{logical: default False. If False, data matrix A and B will not be centered}

\item{scale}{logical: default False. If True, will center and scale, regardless of what center variable is set to}

\item{...}{other parameters to pass in to bisection(...) and bisection.multiple(...)}
}
\value{
values(eigenvalues), vectors (eigenvectors), tau (Lagrange Multiplier) associated with unique component analysis
}
\description{
Run unique component analysis
}
